% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriverWGS.R
\name{prepare_elements}
\alias{prepare_elements}
\title{Preparation of elements and active sites for ActiveDriverWGS analysis with the \code{\link{find_drivers}} function}
\usage{
prepare_elements(elements_file, sites_file, window_size = 50000,
  mc.cores = 1)
}
\arguments{
\item{elements_file}{path to a BED12 file of the elements to analyze}

\item{sites_file}{path to a BED4 file of active sites to use in the analysis}

\item{window_size}{numeric value for the size of the background sequence window to use on each side of the given elements. Default value is 50000.}

\item{mc.cores}{numeric indicating the number of computing cores to use. Default value is 1.}
}
\value{
list containing element and active site information needed for ActiveDriverWGS analysis
}
\description{
Preparation of elements and active sites for ActiveDriverWGS analysis with the \code{\link{find_drivers}} function
}
